Co-locate sentry client config
