+++ packages/observability/keys.ts
@@ -11,11 +11,15 @@ export const keys = () =>
       SENTRY_ORG: z.string().min(1).optional(),
       SENTRY_PROJECT: z.string().min(1).optional(),
     },
+    client: {
+      // Added by Sentry Integration, Vercel Marketplace
+      NEXT_PUBLIC_SENTRY_DSN: z.string().min(1).url().optional(),
+    },
     runtimeEnv: {
       BETTERSTACK_API_KEY: process.env.BETTERSTACK_API_KEY,
       BETTERSTACK_URL: process.env.BETTERSTACK_URL,
-
       SENTRY_ORG: process.env.SENTRY_ORG,
       SENTRY_PROJECT: process.env.SENTRY_PROJECT,
+      NEXT_PUBLIC_SENTRY_DSN: process.env.NEXT_PUBLIC_SENTRY_DSN,
     },
   });
